[New] Add unit test & rename some keywords & move can_utils & fix can_fd & update common_protocol methods 1. Add unit test for common_protocol 2. Rename keyword:"device" to "protocol" (common_device -> common_protocol) 3. Move can_utils from cyberdog_utils to common_protocol 4. Do not depend cyberdog_utils any more 5. Change cmake project name "common_protocol" to "cyberdog_common_protocol" 6. Fix can_op_=null when using can_fd mode(enable_fd=true) 7. Add class state_collector to get all error state 8. Change for_send flag to class protocol constructor
