--- cyberdog_common/common_device/CMakeLists.txt
@@ -1,31 +1,0 @@
-cmake_minimum_required(VERSION 3.8)
-project(common_device)
-
-if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
-  add_compile_options(-Wall -Wextra -Wpedantic)
-endif()
-
-# find dependencies
-find_package(ament_cmake REQUIRED)
-find_package(toml11_vendor REQUIRED)
-find_package(cyberdog_utils REQUIRED)
-
-set(dependencies
-  toml11_vendor
-  cyberdog_utils
-)
-
-install(
-  DIRECTORY include/
-  DESTINATION include/
-)
-
-if(BUILD_TESTING)
-  find_package(ament_lint_auto REQUIRED)
-  ament_lint_auto_find_test_dependencies()
-  find_package(ament_cmake_gtest REQUIRED)
-endif()
-
-ament_export_include_directories(include)
-ament_export_dependencies(${dependencies})
-ament_package()
