--- cyberdog_common/common_device/package.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>common_device</name>
-  <version>1.0.0</version>
-  <description>Easier to use peripherals with toml description file</description>
-  <maintainer email="zhengjunyuan1@xiaomi.com">ZhengJunyuan</maintainer>
-  <license>Apache License, Version 2.0</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>toml11_vendor</depend>
-  <depend>cyberdog_utils</depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
