+++ cyberdog_common/common_protocol/test/common_protocol_test/parser/can/initTest_failed_3.toml
@@ -1,0 +1,224 @@
+# Testing var error
+protocol = "can"
+name = "initTest_failed_3"
+
+can_interface = "can0"
+extended_frame = false
+canfd_enable = false
+timeout_us = 1234567
+
+[[var]]
+description = ""
+can_id = "0x0300 00 00"
+var_name = "bool_var"
+var_type = "bool"
+parser_param = [0, 1, 1]
+
+[[var]]
+description = ""
+can_id = "0x030"
+var_name = "bool_var"
+var_type = "bool"
+parser_param = [0, 3, 0]
+
+[[var]]
+description = ""
+can_id = "0x031"
+var_name = "bool_var"
+var_type = "bool"
+parser_param = [0, 6, 3]
+
+[[var]]
+description = ""
+can_id = "0x031"
+var_name = "bool_var_1"
+var_type = "haha"
+parser_param = [0, 6, 3]
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_2"
+var_type = "bool"
+parser_param = [0, 1, 3]
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_3"
+var_type = "bool"
+parser_param = [1, 8, 3]
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_4"
+var_type = "bool"
+parser_param = [1, 3, 8]
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_5"
+var_type = "bool"
+parser_param = [8, 3, 0]
+
+[[var]]
+description = ""
+can_id = "032"
+var_name = "bool_var_6"
+var_type = "bool"
+parser_param = [-8, 3, 0]
+
+[[var]]
+description = ""
+can_id = "0x0k2"
+var_name = "bool_var_7"
+var_type = "bool"
+parser_param = [2, -3, 0]
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_8"
+var_type = "bool"
+parser_param = [2, 0, -1]
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_9"
+var_type = "bool"
+parser_param = ['$', '$', '$']
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_10"
+var_type = "bool"
+parser_param = ["asd", "123", "52"]
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_11"
+var_type = "bool"
+parser_param = []
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_12"
+var_type = "bool"
+parser_type = "bit"
+parser_param = [1, 2]
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_13"
+var_type = "bool"
+parser_type = "bit"
+parser_param = "[1, 2]"
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_14"
+var_type = "bool"
+parser_type = "bit"
+parser_param = [1, 2, 1, 5]
+
+[[var]]
+description = ""
+can_id = "0x032"
+var_name = "bool_var_15"
+var_type = "bool"
+parser_param = [1, 2, 1, 5]
+
+[[var]]
+description = ""
+can_id = "0x3FF"
+var_name = "u8_var"
+var_type = "u8"
+parser_type = "var"
+parser_param = [0, 0]
+
+[[var]]
+description = ""
+can_id = "0x3FF"
+var_name = "u16_var"
+var_type = "u16"
+parser_type = "var"
+parser_param = [0, 1]
+
+[[var]]
+description = ""
+can_id = "0x3FF"
+var_name = "u32_var"
+var_type = "u32"
+parser_type = "var"
+parser_param = [0, 3]
+
+[[var]]
+description = ""
+can_id = "0x3FF"
+var_name = "u8_1_bit"
+var_type = "u8"
+parser_type = "bit"
+parser_param = [0, 3, 3]
+
+[[var]]
+description = ""
+can_id = "0x3FF"
+var_name = "u8_4_bit"
+var_type = "u8"
+parser_type = "bit"
+parser_param = [0, 5, 2]
+
+[[var]]
+description = ""
+can_id = "0x3FF"
+var_name = "u8_4_bit"
+var_type = "u8"
+parser_type = "bit"
+parser_param = [0, 5, 2]
+
+[[var]]
+description = ""
+can_id = "0x3FF"
+var_name = "u8_var_0"
+var_type = "u8"
+parser_type = "hahaha"
+parser_param = [0, 5, 2]
+
+[[var]]
+description = ""
+can_id = "0x3FF"
+var_name = "u8_var_1"
+var_type = "u8"
+parser_type = "var"
+parser_param = [5, 0]
+
+[[var]]
+description = ""
+can_id = "0x3FF"
+var_name = "u8_var_2"
+var_type = 155
+parser_type = "var"
+parser_param = [105, 105]
+
+[[var]]
+description = ""
+can_id = "0x3FF"
+var_name = "u8_var_3"
+var_type = "u8"
+parser_type = "var"
+parser_param = [0, 0, 5]
+
+[[var]]
+description = ""
+can_id = "0x3FF"
+var_name = "u8_var_4"
+var_type = "u8"
+parser_param = [0, 0, 5, 5]
